// front/client/types.gen.ts
export type ModelsNode = {
    created_at?: string;
    id?: string;
    mime?: string;
    name?: string;
    owner_id?: string;
    /**
     * root: empty
     */
    parent_id?: string;
    /**
     * local storage path for files
     */
    path?: string;
    /**
     * bytes for files
     */
    size?: number;
    /**
     * "file" | "folder"
     */
    type?: string;
    updated_at?: string;
};